+++ resources/views/powers/index.blade.php
@@ -13,18 +13,26 @@
     <div class="container-fluid">
         @if(session()->has('success'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('success')}}
             </div>
         @elseif(session()->has('updated'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('updated')}}
             </div>
         @elseif(session()->has('deleted'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('deleted')}}
             </div>
         @elseif(session()->has('connection'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('connection')}}
             </div>
         @endif
@@ -36,7 +44,6 @@
                                placeholder="Search By Generator Id, Type Or Capacity"
                                value="{{request()->query('search')}}">
                         <div class="input-group-addon">
-                            {{--                            <span class="input-group-text"><i class="fas fa-search"></i></span>--}}
                             <button id="search" type="button" class="btn btn-primary">
                                 <i class="fas fa-search"></i>
                             </button>
@@ -45,20 +52,20 @@
                 </div>
             </div>
             <div class="col">
-                @canany(['site-create','site-edit','site-delete'])
+                @can('site-create')
                     <div class="text-right">
                         <a href="{{route('powers.create')}}"
-                           class="btn btn-outline-primary mb-2"><i
+                           class="btn btn-outline-dark mb-2"><i
                                 class="fas fa-plus-square fa-2x"></i></a>
                     </div>
-                @endcanany
+                @endcan
             </div>
         </div>
-        <div class="card border-success mb-3">
-            <div class="card-header bg-gradient-primary font-weight-bold">Generators</div>
+        <div class="card border-dark mb-3">
+            <div class="card-header bg-gradient-gray-dark font-weight-bold">Generators</div>
             <div class="card-body text-black-50">
                 @if($powers->isNotEmpty())
-                    <table class="table table-bordered table-responsive">
+                    <table class="table table-bordered table-responsive border-primary">
                         <thead>
                         <tr class="bg-gradient-primary">
                             <th scope="col">Id</th>
@@ -87,10 +94,12 @@
                             <th scope="col">Work Order Id</th>
                             <th scope="col">Created At</th>
                             <th scope="col">Updated At</th>
-                            @canany(['site-create','site-edit','site-delete'])
+                            @can('site-edit')
                                 <th scope="col">Update</th>
+                            @endcan
+                            @can('site-delete')
                                 <th scope="col">Delete</th>
-                            @endcanany
+                            @endcan
                         </tr>
                         </thead>
                         <tbody>
@@ -121,16 +130,18 @@
                                 <td>{{$power->work_order_id}}</td>
                                 <td>{{ $power->created_at->format('Y-m-d') }}</td>
                                 <td>{{ $power->updated_at->format('Y-m-d') }}</td>
-                                @canany(['site-create','site-edit','site-delete'])
+                                @can('site-edit')
                                     <td><a href="{{route('powers.edit', $power->id)}}"
                                            class="btn btn-primary btn-sm">Update</a></td>
+                                @endcan
+                                @can('site-delete')
                                     <td>
                                         <button class="btn btn-danger btn-sm"
                                                 onclick="handleDelete({{$power->id}})">
                                             Delete
                                         </button>
                                     </td>
-                                @endcanany
+                                @endcan
                             </tr>
                         @endforeach
                         </tbody>
