+++ resources/views/rectifiers/index.blade.php
@@ -14,18 +14,26 @@
     <div class="container-fluid">
         @if(session()->has('success'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('success')}}
             </div>
         @elseif(session()->has('updated'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('updated')}}
             </div>
         @elseif(session()->has('deleted'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('deleted')}}
             </div>
         @elseif(session()->has('connection'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('connection')}}
             </div>
         @endif
@@ -46,20 +54,20 @@
                 </div>
             </div>
             <div class="col">
-                @canany(['site-create','site-edit','site-delete'])
+                @can('site-create')
                     <div class="text-right">
-                        <a href="{{route('rectifiers.create')}}" class="btn btn-outline-primary mb-2"><i
+                        <a href="{{route('rectifiers.create')}}" class="btn btn-outline-dark mb-2"><i
                                 class="fas fa-plus-square fa-2x"></i></a>
                     </div>
-                @endcanany
+                @endcan
             </div>
         </div>
     </div>
-    <div class="card border-success mb-3">
-        <div class="card-header bg-gradient-primary font-weight-bold">Rectifier</div>
+    <div class="card border-dark mb-3">
+        <div class="card-header bg-gradient-gray-dark font-weight-bold">Rectifier</div>
         <div class="card-body text-black-50">
             @if($rectifiers->isNotEmpty())
-                <table class="table table-bordered table-responsive">
+                <table class="table table-bordered table-responsive border-primary">
                     <thead>
                     <tr class="bg-gradient-primary">
                         <th scope="col">Id</th>
@@ -80,10 +88,12 @@
                         <th scope="col">Work Order Id</th>
                         <th scope="col">Created At</th>
                         <th scope="col">Updated At</th>
-                        @canany(['site-create','site-edit','site-delete'])
+                        @can('site-edit')
                             <th scope="col">Update</th>
+                        @endcan
+                        @can('site-delete')
                             <th scope="col">Delete</th>
-                        @endcanany
+                        @endcan
                     </tr>
                     </thead>
                     <tbody>
@@ -107,15 +117,17 @@
                             <td>{{$rectifier->work_order_id}}</td>
                             <td>{{ $rectifier->created_at->format('Y-m-d') }}</td>
                             <td>{{ $rectifier->updated_at->format('Y-m-d') }}</td>
-                            @canany(['site-create','site-edit','site-delete'])
+                            @can('site-edit')
                                 <td><a href="{{route('rectifiers.edit', $rectifier->id)}}"
                                        class="btn btn-primary btn-sm">Update</a></td>
+                            @endcan
+                            @can('site-delete')
                                 <td>
                                     <button class="btn btn-danger btn-sm" onclick="handleDelete({{$rectifier->id}})">
                                         Delete
                                     </button>
                                 </td>
-                            @endcanany
+                            @endcan
                         </tr>
                     @endforeach
                     </tbody>
