+++ resources/views/ups/index.blade.php
@@ -13,18 +13,26 @@
     <div class="container-fluid">
         @if(session()->has('success'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('success')}}
             </div>
         @elseif(session()->has('updated'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('updated')}}
             </div>
         @elseif(session()->has('deleted'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('deleted')}}
             </div>
         @elseif(session()->has('connection'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('connection')}}
             </div>
         @endif
@@ -32,7 +40,8 @@
             <div class="col-md-4 col-xl-4 mb-3">
                 <div class="sidebar px-4 py-md-0">
                     <form action="{{route('ups.index')}}" class="input-group" method="get">
-                        <input type="text" class="form-control" name="search" placeholder="Search By Ups Id, Type Or Model">
+                        <input type="text" class="form-control" name="search"
+                               placeholder="Search By Ups Id, Type Or Model">
                         <div class="input-group-addon">
                             <button id="search" type="button" class="btn btn-primary">
                                 <i class="fas fa-search"></i>
@@ -42,19 +51,19 @@
                 </div>
             </div>
             <div class="col">
-                @canany(['site-create','site-edit','site-delete'])
+                @can('site-create')
                     <div class="text-right">
-                        <a href="{{route('ups.create')}}" class="btn btn-outline-primary mb-2"><i
+                        <a href="{{route('ups.create')}}" class="btn btn-outline-dark mb-2"><i
                                 class="fas fa-plus-square fa-2x"></i></a>
                     </div>
-                @endcanany
+                @endcan
             </div>
         </div>
-        <div class="card border-success mb-3">
-            <div class="card-header bg-gradient-primary font-weight-bold">Ups</div>
+        <div class="card border-dark mb-3">
+            <div class="card-header bg-gradient-gray-dark font-weight-bold">Ups</div>
             <div class="card-body text-black-50">
                 @if($upss->isNotEmpty())
-                    <table class="table table-bordered table-responsive">
+                    <table class="table table-bordered table-responsive border-primary">
                         <thead>
                         <tr class="bg-gradient-primary">
                             <th scope="col">Id</th>
@@ -71,12 +80,15 @@
                             <th scope="col">LLD Number</th>
                             <th scope="col">Commission Date</th>
                             <th scope="col">Site Id</th>
+                            <th scope="col">Work Order Id</th>
                             <th scope="col">Created At</th>
                             <th scope="col">Updated At</th>
-                            @canany(['site-create','site-edit','site-delete'])
+                            @can('site-edit')
                                 <th scope="col">Update</th>
+                            @endcan
+                            @can('site-delete')
                                 <th scope="col">Delete</th>
-                            @endcanany
+                            @endcan
                         </tr>
                         </thead>
                         <tbody>
@@ -96,18 +108,21 @@
                                 <td>{{ $ups->lld_number }}</td>
                                 <td>{{ $ups->commission_date }}</td>
                                 <td>{{ $ups->site_id }}</td>
+                                <td>{{$ups->work_order_id}}</td>
                                 <td>{{ $ups->created_at->format('Y-m-d') }}</td>
                                 <td>{{ $ups->updated_at->format('Y-m-d') }}</td>
-
-                                @canany(['site-create','site-edit','site-delete'])
+                                @can('site-edit')
                                     <td><a href="{{route('ups.edit', $ups->id)}}"
-                                           class="btn btn-primary btn-sm">Update</a></td>
+                                           class="btn btn-primary btn-sm">Update</a>
+                                    </td>
+                                @endcan
+                                @can('site-delete')
                                     <td>
                                         <button class="btn btn-danger btn-sm" onclick="handleDelete({{$ups->id}})">
                                             Delete
                                         </button>
                                     </td>
-                                @endcanany
+                                @endcan
                             </tr>
                         @endforeach
                         </tbody>
