+++ resources/views/workorders/index.blade.php
@@ -9,22 +9,29 @@
 {{--@endsection--}}
 
 @section('content')
-
     <div class="container-fluid">
         @if(session()->has('success'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('success')}}
             </div>
         @elseif(session()->has('updated'))
             <div class="alert alert-success">
+                <span style="font-size: 2em; color: #00a87d">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('updated')}}
             </div>
         @elseif(session()->has('deleted'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('deleted')}}
             </div>
         @elseif(session()->has('connection'))
             <div class="alert alert-danger">
+                <span style="font-size: 2em; color: #ff0000">
+                    <i class="fas fa-info-circle"></i></span>
                 {{session()->get('connection')}}
             </div>
         @endif
@@ -32,7 +39,8 @@
             <div class="col-md-4 col-xl-4 mb-3">
                 <div class="sidebar px-4 py-md-0">
                     <form action="{{route('workorders.index')}}" class="input-group" method="get">
-                        <input type="text" class="form-control" name="search" placeholder="Search By Work Order Id Or Number">
+                        <input type="text" class="form-control" name="search"
+                               placeholder="Search By Work Order Id Or Number">
                         <div class="input-group-addon">
                             <button id="search" type="button" class="btn btn-primary">
                                 <i class="fas fa-search"></i>
@@ -42,16 +50,16 @@
                 </div>
             </div>
             <div class="col">
-                @canany(['site-create','site-edit','site-delete'])
+                @can('site-create')
                     <div class="text-right">
-                        <a href="{{route('workorders.create')}}" class="btn btn-outline-primary mb-2"><i
+                        <a href="{{route('workorders.create')}}" class="btn btn-outline-dark mb-2"><i
                                 class="fas fa-plus-square fa-2x"></i></a>
                     </div>
-                @endcanany
+                @endcan
             </div>
         </div>
-        <div class="card border-success mb-3">
-            <div class="card-header bg-gradient-primary font-weight-bold">Work Orders</div>
+        <div class="card border-dark mb-3">
+            <div class="card-header bg-gradient-gray-dark font-weight-bold">Work Orders</div>
             <div class="card-body text-black-50">
                 @if($workorders->isNotEmpty())
                     <table class="table table-bordered">
@@ -59,13 +67,15 @@
                         <tr class="bg-gradient-primary">
                             <th scope="col">Id</th>
                             <th scope="col">Work Order Number</th>
-                            <th scope="col">Site Id</th>
                             <th scope="col">Created At</th>
                             <th scope="col">Updated At</th>
-                            @canany(['site-create','site-edit','site-delete'])
+                            <th scope="col">Details</th>
+                            @can('site-edit')
                                 <th scope="col">Update</th>
+                            @endcan
+                            @can('site-delete')
                                 <th scope="col">Delete</th>
-                            @endcanany
+                            @endcan
                         </tr>
                         </thead>
                         <tbody>
@@ -73,19 +83,25 @@
                             <tr>
                                 <th scope="row">{{ $workorder->id }}</th>
                                 <td>{{ $workorder->work_orders_number }}</td>
-                                <td>{{ $workorder->site_id }}</td>
                                 <td>{{ $workorder->created_at->format('Y-m-d') }}</td>
                                 <td>{{ $workorder->updated_at->format('Y-m-d') }}</td>
-                                @canany(['site-create','site-edit','site-delete'])
+                                <td>
+                                    <a href="{{route('workorders.show', $workorder->id)}}"
+                                       class="btn btn-success btn-sm">Detail</a>
+                                </td>
+                                @can('site-edit')
                                     <td><a href="{{route('workorders.edit', $workorder->id)}}"
-                                           class="btn btn-primary btn-sm">Update</a></td>
+                                           class="btn btn-primary btn-sm">Update</a>
+                                    </td>
+                                @endcan
+                                @can('site-delete')
                                     <td>
                                         <button class="btn btn-danger btn-sm"
                                                 onclick="handleDelete({{$workorder->id}})">
                                             Delete
                                         </button>
                                     </td>
-                                @endcanany
+                                @endcan
                             </tr>
                         @endforeach
                         </tbody>
